Rename 'id' => 'name' (socket.id), player destruction through server (was front only before that)

+++ front/js/Racer.js
@@ -64,6 +64,7 @@ Racer.prototype.playerUpdate = function(){
 
   // Player's elements collisions
   this.physics.arcade.collide(this.obstacles, player, function(){
+    _this.player.data.destroy(_this.player.data.name);
     _this.destroyPlayer(_this.player.el, _this.player.data.color);
   });
   if(player.explosion)
@@ -107,9 +108,6 @@ Racer.prototype.foesUpdate = function(){
     foe.trail.x = foe.position.x;
     foe.trail.y = foe.position.y + foe.height / 2;
     // Foes collisions
-    this.physics.arcade.collide(this.obstacles, foe, function(){
-      _this.destroyPlayer(foe, foe.color);
-    });
     if(foe.explosion)
       this.physics.arcade.collide(this.obstacles, foe.explosion);
     if(foe.trail)
@@ -117,13 +115,13 @@ Racer.prototype.foesUpdate = function(){
   }
 };
 
-Racer.prototype.createFoe = function(color, id, pos){
+Racer.prototype.createFoe = function(color, name, pos){
   // Create foe
   var foe = this.foes.create(0, 0, this.playerCreator(this.player.data.size, color));
   // Add trail to foe
   this.addTrail(foe, color);
   // Add name
-  foe.name = id;
+  foe.name = name;
   // Add color
   foe.color = color;
   // Set pos to 0,0 if foe just arrive
@@ -138,20 +136,20 @@ Racer.prototype.createFoe = function(color, id, pos){
   }
 };
 
-Racer.prototype.updateFoe = function(pos, id){
+Racer.prototype.updateFoe = function(pos, name){
   // Update foe's position when he's moving
   for(var i = 0; i < this.foes.children.length; i++){
-    if(this.foes.children[i].name == id){
+    if(this.foes.children[i].name == name){
       this.foes.children[i].x = pos.x;
       this.foes.children[i].y = pos.y;
     }
   }
 };
 
-Racer.prototype.destroyFoe = function(id){
+Racer.prototype.destroyFoe = function(name){
   // Destroy foe on disconnect
   for(var i = 0; i < this.foes.children.length; i++){
-    if(this.foes.children[i].name == id){
+    if(this.foes.children[i].name == name){
       this.destroyPlayer(this.foes.children[i], this.foes.children[i].color);
     }
   }
@@ -170,9 +168,11 @@ Racer.prototype.playerCreator = function(size, color){
 
 Racer.prototype.destroyPlayer = function(player, color){
   // Player kill + destruction animation
-  this.createExplosion(player, color);
-  player.kill();
-  player.trail.kill();
+  if(player.alive){
+    this.createExplosion(player, color);
+    player.kill();
+    player.trail.kill();
+  }
 };
 
 Racer.prototype.createExplosion = function(player, color){
